Several types of changes:
1) tabs expanded to spaces
2) more general updates to javadocs
3) Class method convertEnumValueToName -- added more error checking. And, after
talking with Peter this should be moved to the datatype class most likely. But,
check in a copy of what we have for now.
4) Class method copyAttributes(HObject, HObject)- more error checking. more doc
and, made static final - not just static.
5) Class method setObjectName-- this is problematic.   See the source for lots
 of comments about problems.    I added more documentation & error checking, but
it doesn't mesh well with what I think is supposed to be going on.  I don't 
think having a / in the new name makes sense.  I think if you call this directly
you aren't going to get the in-memory copy of the name updated.    
Peter will review.    

Basically, wanted to check point these and move on.   But, definitely not done.